+++ src/nexgen/nxs_write/NXclassWriters.py
@@ -9,7 +9,6 @@ from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
 import numpy as np
-from hdf5plugin import Bitshuffle
 
 from .. import (
     get_iso_timestamp,
@@ -19,7 +18,10 @@ from .. import (
     units_of_time,
     ureg,
 )
-from . import calculate_origin, create_attributes, set_dependency
+from . import calculate_origin, create_attributes, set_dependency, write_compressed_copy
+
+# from hdf5plugin import Bitshuffle   # noqa: F401
+
 
 import h5py  # isort: skip
 
@@ -392,55 +394,55 @@ def write_NXsource(nxsfile: h5py.File, source: Dict):
     nxsource.create_dataset("type", data=np.string_(source["type"]))
 
 
-# Copy and compress mask or flatfield file in NXdetector
-def write_compressed_copy(
-    nxgroup: h5py.Group,
-    dset_name: str,
-    data: np.ndarray = None,
-    filename: Path | str = None,
-    dset_key: str = None,
-    block_size: int = 0,
-):
-    """
-    Write a compressed copy of some dataset in the desired HDF5 group, using the Bitshuffle filter with lz4 compression.
-    The main application for this function in nexgen is to write a compressed copy of a pixel mask or a flatfield file/dataset \
-    directly into the NXdetector group of a NXmx NeXus file.
-    The data and filename arguments are mutually exclusive as only one of them can be used as input.
-    If a filename is passed, it is also required to pass the key for the relevant dataset to be copied. Failure to do so will result \
-    in nothing being written to the NeXus file.
+# # Copy and compress mask or flatfield file in NXdetector
+# def write_compressed_copy(
+#     nxgroup: h5py.Group,
+#     dset_name: str,
+#     data: np.ndarray = None,
+#     filename: Path | str = None,
+#     dset_key: str = None,
+#     block_size: int = 0,
+# ):
+#     """
+#     Write a compressed copy of some dataset in the desired HDF5 group, using the Bitshuffle filter with lz4 compression.
+#     The main application for this function in nexgen is to write a compressed copy of a pixel mask or a flatfield file/dataset \
+#     directly into the NXdetector group of a NXmx NeXus file.
+#     The data and filename arguments are mutually exclusive as only one of them can be used as input.
+#     If a filename is passed, it is also required to pass the key for the relevant dataset to be copied. Failure to do so will result \
+#     in nothing being written to the NeXus file.
 
-    Args:
-        nxgroup (h5py.Group): Handle to HDF5 group.
-        dset_name (str): Name of the new dataset to be written.
-        data (np.ndarray, optional): Dataset to be compressed. Defaults to None.
-        filename (Path | str, optional): Filename containing the dataset to be compressed into the NeXus file. Defaults to None.
-        dset_key (str, optional): Dataset name inside the passed file. Defaults to None.
-        block_size (int, optional): Number of elements per block, it needs to be divisible by 8. Defaults to 0.
+#     Args:
+#         nxgroup (h5py.Group): Handle to HDF5 group.
+#         dset_name (str): Name of the new dataset to be written.
+#         data (np.ndarray, optional): Dataset to be compressed. Defaults to None.
+#         filename (Path | str, optional): Filename containing the dataset to be compressed into the NeXus file. Defaults to None.
+#         dset_key (str, optional): Dataset name inside the passed file. Defaults to None.
+#         block_size (int, optional): Number of elements per block, it needs to be divisible by 8. Defaults to 0.
 
-    Raises:
-        ValueError: If both a dataset and a filename have been passed to the function.
-    """
-    if data is not None and filename is not None:
-        raise ValueError(
-            "The dset and filename arguments are mutually exclusive."
-            "Please pass only the one from which the data should be copied."
-        )
-    if filename and not dset_key:
-        NXclass_logger.warning(
-            f"Missing key to find the dataset to be copied inside {filename}. {dset_name} will not be written into the NeXus file."
-        )
-        return
+#     Raises:
+#         ValueError: If both a dataset and a filename have been passed to the function.
+#     """
+#     if data is not None and filename is not None:
+#         raise ValueError(
+#             "The dset and filename arguments are mutually exclusive."
+#             "Please pass only the one from which the data should be copied."
+#         )
+#     if filename and not dset_key:
+#         NXclass_logger.warning(
+#             f"Missing key to find the dataset to be copied inside {filename}. {dset_name} will not be written into the NeXus file."
+#         )
+#         return
 
-    if filename:
-        with h5py.File(filename, "r") as fh:
-            data = fh[dset_key][()]
+#     if filename:
+#         with h5py.File(filename, "r") as fh:
+#             data = fh[dset_key][()]
 
-    nxgroup.create_dataset(
-        dset_name, data=data, **Bitshuffle(nelems=block_size, lz4=True)
-    )
-    NXclass_logger.info(
-        f"A compressed copy of the {dset_name} has been written into the NeXus file."
-    )
+#     nxgroup.create_dataset(
+#         dset_name, data=data, **Bitshuffle(nelems=block_size, lz4=True)
+#     )
+#     NXclass_logger.info(
+#         f"A compressed copy of the {dset_name} has been written into the NeXus file."
+#     )
 
 
 # NXdetector writer
